fixed folder structure by destructuring components
