+++ src/App.jsx
@@ -1,64 +1,14 @@
-import { useState } from "react";
-import { Header, CoreConcept, TabButton } from "@components";
-import { CORE_CONCEPTS, EXAMPLES } from "./data";
+import { Header, CoreConcepts, Examples } from "@components";
 
 function App() {
-	const [selectedTopic, setSelectedTopic] = useState("components");
-
-	function handleSelect(selectedButton) {
-		setSelectedTopic(selectedButton);
-	}
-
 	return (
-		<div>
+		<>
 			<Header />
 			<main>
-				<section id="core-concepts">
-					<h2>Core Concepts</h2>
-					<ul>
-						{CORE_CONCEPTS.map((conceptItem) => (
-							<CoreConcept key={conceptItem.id} {...conceptItem} />
-						))}
-					</ul>
-				</section>
-				<section id="examples">
-					<h2>Examples</h2>
-					<menu>
-						<TabButton
-							isSelected={selectedTopic === "components"}
-							onClick={() => handleSelect("components")}
-						>
-							Components
-						</TabButton>
-						<TabButton
-							isSelected={selectedTopic === "jsx"}
-							onClick={() => handleSelect("jsx")}
-						>
-							JSX
-						</TabButton>
-						<TabButton
-							isSelected={selectedTopic === "props"}
-							onClick={() => handleSelect("props")}
-						>
-							Props
-						</TabButton>
-						<TabButton
-							isSelected={selectedTopic === "state"}
-							onClick={() => handleSelect("state")}
-						>
-							State
-						</TabButton>
-					</menu>
-					<div id="tab-content">
-						<h3>{EXAMPLES[selectedTopic].title}</h3>
-						<p>{EXAMPLES[selectedTopic].description}</p>
-						<pre>
-							<code>{EXAMPLES[selectedTopic].code}</code>
-						</pre>
-					</div>
-				</section>
+				<CoreConcepts />
+				<Examples />
 			</main>
-		</div>
+		</>
 	);
 }
