+++ src/components/index.js
@@ -1,3 +1,6 @@
 export { Header } from "./Header";
 export { CoreConcept } from "./CoreConcept";
 export { TabButton } from "./TabButton";
+export { CoreConcepts } from "./CoreConcepts";
+export { Examples } from "./Examples";
+export { Section } from "./Section";
